% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_reactibble}
\alias{as_reactibble}
\title{Convert to a reactibble object}
\usage{
as_reactibble(
  x,
  ...,
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal"),
  rownames = pkgconfig::get_config("tibble::rownames", NULL)
)
}
\arguments{
\item{x}{forwarded to tibble::as_tibble}

\item{...}{forwarded to tibble::as_tibble}

\item{.rows}{forwarded to tibble::as_tibble}

\item{.name_repair}{forwarded to tibble::as_tibble}

\item{rownames}{forwarded to tibble::as_tibble}
}
\description{
Convert to a reactibble object
}
