% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{as_tibble.reactibble}
\alias{as_tibble.reactibble}
\title{Convert to tibble}
\usage{
as_tibble.reactibble(
  x,
  ...,
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal"),
  rownames = pkgconfig::get_config("tibble::rownames", NULL)
)
}
\arguments{
\item{x}{react tibble object}

\item{...}{forwarded to tibble::as_tibble}

\item{.rows}{forwarded to tibble::as_tibble}

\item{.name_repair}{Treatment of problematic column names}

\item{rownames}{rownames How to treat existing row names of a data frame or matrix}
}
\description{
Convert to tibble
}
