% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_reactibble}
\alias{is_reactibble}
\title{Convert to a reactibble object}
\usage{
is_reactibble(x)
}
\arguments{
\item{x}{A data frame, list, matrix, or other object that could reasonably be coerced to a tibble.}
}
\description{
Convert to a reactibble object
}
