% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{materialize}
\alias{materialize}
\title{convert reactive columns to static columns}
\usage{
materialize(x, ...)
}
\arguments{
\item{x}{a reactibble object}

\item{...}{bare column names}
}
\description{
If no column names are provided, the full reactibble is materialized. The
class "reactibble" is preserved.
}
