% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{tbl_sum.reactibble}
\alias{tbl_sum.reactibble}
\alias{vec_ptype_abbr.reactive_col}
\alias{pillar_shaft.reactive_col}
\alias{format.reactive_col}
\title{methods}
\usage{
tbl_sum.reactibble(x)

vec_ptype_abbr.reactive_col(x)

pillar_shaft.reactive_col(x, ...)

\method{format}{reactive_col}(x, ...)
}
\arguments{
\item{x}{Object to summarise.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
methods
}
