% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate.reactibble}
\alias{mutate.reactibble}
\alias{transmute.reactibble}
\title{modify a reactibble object}
\usage{
mutate.reactibble(.data, ...)

transmute.reactibble(.data, ...)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{...}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> Name-value pairs.
The name gives the name of the column in the output.

The value can be:
\itemize{
\item A vector of length 1, which will be recycled to the correct length.
\item A vector the same length as the current group (or the whole data frame
if ungrouped).
\item \code{NULL}, to remove the column.
\item A data frame or tibble, to create multiple columns in the output.
}}
}
\description{
These work exactly like dplyr's mutate and transmute, except that one can
define reactive columns using \code{~}.
}
