% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{pillar_methods}
\alias{pillar_methods}
\alias{pillar_shaft.reactive_col}
\title{pillar methods}
\usage{
pillar_shaft.reactive_col(x, ...)
}
\arguments{
\item{x}{A vector to format}

\item{...}{Arguments passed to methods.}
}
\description{
pillar methods
}
