% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reactibble}
\alias{reactibble}
\title{Build a reactibble object}
\usage{
reactibble(...)
}
\arguments{
\item{...}{A set of name-value pairs, use \code{~} to define a reactive column}
}
\description{
Build a reactibble object
}
