% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{rt_bind_rows}
\alias{rt_bind_rows}
\title{Efficiently bind multiple data frames by row and column}
\usage{
rt_bind_rows(..., .id = NULL)
}
\arguments{
\item{...}{Data frames to combine.

Each argument can either be a data frame, a list that could be a data
frame, or a list of data frames.

When row-binding, columns are matched by name, and any missing
columns will be filled with NA.

When column-binding, rows are matched by position, so all data
frames must have the same number of rows. To match by value, not
position, see \link[dplyr]{mutate-joins}.}

\item{.id}{Data frame identifier.

When \code{.id} is supplied, a new column of identifiers is
created to link each row to its original data frame. The labels
are taken from the named arguments to \code{bind_rows()}. When a
list of data frames is supplied, the labels are taken from the
names of the list. If no names are found a numeric sequence is
used instead.}
}
\description{
Counterpart of \code{dplyr::bind_rows} that works efficiently on \emph{"reactibble"}
objects. While \code{bind_rows()} can be used on "reactibbles" (at time of writing),
it is brittle and inefficient, as it triggers more refreshes than necessary.
}
