% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_reactibble}
\alias{as_reactibble}
\title{Convert to a reactibble object}
\usage{
as_reactibble(x)
}
\arguments{
\item{x}{A data frame, list, matrix, or other object that could reasonably be coerced to a tibble.}
}
\description{
Convert to a reactibble object
}
